MaJ : b1

+++ modules/AlgoPacking/src/MultiDArrayObject.cxx
@@ -8,11 +8,10 @@ namespace merope {
 
 vector<array<size_t, 2>> vox::auxi::getSegmentsFromPeriodic(const array<long, 2>& perLimits, size_t nbNode) {
     assert(perLimits[0] <= perLimits[1]);
-    //
     if (perLimits[1] - perLimits[0] >= nbNode) {
         return vector<array<size_t, 2>>{array<size_t, 2>{0, nbNode}};
     } else {
-        size_t k = 0; // for perLimits[0] - k * nbNode
+        size_t k = 0;  // for perLimits[0] - k * nbNode
         long i0_long = perLimits[0];
         while (i0_long < 0) {
             k--;
@@ -22,8 +21,8 @@ vector<array<size_t, 2>> vox::auxi::getSegmentsFromPeriodic(const array<long, 2>
             k++;
             i0_long -= nbNode;
         }
-        size_t i0 = i0_long;
-        size_t i1 = perLimits[1] - k * nbNode;
+        size_t i0 = static_cast<size_t>(i0_long);
+        size_t i1 = static_cast<size_t>(perLimits[1] - k * nbNode);
         if (i1 <= nbNode) {
             return vector<array<size_t, 2>>{array<size_t, 2>{i0, i1}};
         } else {
@@ -32,4 +31,4 @@ vector<array<size_t, 2>> vox::auxi::getSegmentsFromPeriodic(const array<long, 2>
     }
 }
 
-} // namespace merope
+}  // namespace merope
